+++ .gitignore
@@ -15,3 +15,15 @@ yarn-error.log
 .yarnrc.yml
 
 .pnp*
+
+
+# Cooler-Env secret key
+test_dir/test.key
+
+
+# Cooler-Env secret key
+test_dir/test.key
+
+
+# Cooler-Env secret key
+test_dir/test.key
