+++ jest.config.js
@@ -1,0 +1,9 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} **/
+module.exports = {
+  testEnvironment: "node",
+  transform: {
+    "^.+.ts?$": ["ts-jest", {}],
+    "^.+\\.(js)$": "babel-jest",
+  },
+  transformIgnorePatterns: ["node_modules/(?!(chalk)/)"],
+};
