+++ package.json
@@ -10,7 +10,7 @@
     "url": "https://github.com/arsood/Cooler-Env.git"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "jest"
   },
   "dependencies": {
     "chalk": "^5.3.0",
@@ -23,20 +23,27 @@
   },
   "bin": "./index.js",
   "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.4",
+    "@types/babel__core": "^7",
+    "@types/babel__preset-env": "^7",
     "@types/chalk": "^2.2.0",
     "@types/clear": "^0.1.4",
     "@types/figlet": "^1.5.8",
     "@types/inquirer": "^9.0.7",
+    "@types/jest": "^29.5.13",
     "@types/minimist": "^1.2.5",
     "@types/node": "^20.11.16",
-    "chai": "^4.3.4",
+    "@types/rimraf": "^4.0.5",
+    "babel-jest": "^29.7.0",
     "eslint": "^7.26.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.2",
     "eslint-plugin-prettier": "^3.4.0",
-    "mocha": "^8.4.0",
+    "jest": "^29.7.0",
     "prettier": "^2.3.0",
-    "rimraf": "^3.0.2"
+    "rimraf": "^6.0.1",
+    "ts-jest": "^29.2.5"
   },
   "packageManager": "yarn@4.2.2"
 }
